More separation of shell and photo booth, app does compile now
