+++ lib/app/photo_booth.routes.dart
@@ -1,4 +1,4 @@
-part of 'photobooth.dart';
+part of 'photo_booth.dart';
 
 List<GoRoute> _rootRoutes = [
   _startRoute,
