+++ lib/app/widgets/hotkey_monitor.dart
@@ -5,7 +5,6 @@ import 'package:flutter/services.dart';
 import 'package:go_router/go_router.dart';
 import 'package:loggy/loggy.dart';
 import 'package:momento_booth/extensions/go_router_extension.dart';
-import 'package:momento_booth/main.dart';
 import 'package:momento_booth/managers/_all.dart';
 import 'package:momento_booth/views/manual_collage_screen/manual_collage_screen.dart';
 import 'package:momento_booth/views/start_screen/start_screen.dart';
@@ -26,7 +25,6 @@ class HotkeyMonitor extends StatelessWidget with UiLoggy {
         SingleActivator(LogicalKeyboardKey.keyH, control: control, meta: meta): () => router.go(StartScreen.defaultRoute),
         SingleActivator(LogicalKeyboardKey.keyR, control: control, meta: meta): LiveViewManager.instance.restoreLiveView,
         SingleActivator(LogicalKeyboardKey.keyS, control: control, meta: meta): () {
-          AppState state = context.findAncestorStateOfType<AppState>()!;
           // ignore: invalid_use_of_protected_member
           // state.setState(() => state.settingsOpen = !state.settingsOpen);
           // loggy.debug("Settings ${state.settingsOpen ? "opened" : "closed"}");
